add delete superkey workflow, move all "helper" functions to their own file
